fix
long to bigdecimal changed

+++ src/main/java/com/asg/shipping/exportManifestBl/dto/ExportManifestBlCreateDto.java
@@ -37,7 +37,7 @@ public class ExportManifestBlCreateDto {
     @Size(max = 200, message = "Agent reference must not exceed 200 characters")
     private String agentReference;
 
-    private Long shipperPoid;
+    private BigDecimal shipperPoid;
     private Long shipperAddressPoid;
     private Long consigneePoid;
     private Long consigneeAddressPoid;
